+++ CHANGELOG.md
@@ -6,6 +6,8 @@ FEATURES:
 
 ENHANCEMENTS:
 
+* Changed 'prefix' to 'abbreviation' in json schema and resource parameter to reflect MSFT naming documentation
+
 BUG FIXES:
 
 * Fixed resource documentation