+++ CHANGELOG.md
@@ -1,3 +1,14 @@
-## 0.1.0 (Unreleased)
+## 0.0.5
 
-BACKWARDS INCOMPATIBILITIES / NOTES:
+FEATURES:
+
+* Experiemental support for downloading json naming schema from URL
+
+ENHANCEMENTS:
+
+* Changed 'prefix' to 'abbreviation' in json schema and resource parameter to reflect MSFT naming documentation
+
+BUG FIXES:
+
+* Fixed resource documentation
+